+++ src/api/axios/Auth/authInterfaces.ts
@@ -1,0 +1,4 @@
+export interface IloginData {
+  email: string;
+  password: string;
+}
